mppa: ucode/pcie: remove ETO when no packets are provided
Because we alloc a mOS UC slot before prepping/checking the packet, it happens
that the first backet of the 8 to sent has errors (len > MTU) which causes the process
to not look at the other packets. In this case, we stiil need to trigger the uCode so
the slot can be returned to mOS for later used.
With the current uC, even an empty job caused an EOT which used buffer on the IO
but was not accounted for in the credit.
This patch checks for 1st len == 0 and in that case stop the uC directly without transfers nor EOT

Signed-off-by: Nicolas Morey-Chaisemartin <nmorey@kalray.eu>